Intake Automatiation ++ State machines Working
