+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/CenterStageRobot/commands/IntakeCommand.java
@@ -5,6 +5,7 @@ import com.arcrobotics.ftclib.command.InstantCommand;
 import com.arcrobotics.ftclib.command.SequentialCommandGroup;
 import com.arcrobotics.ftclib.command.WaitCommand;
 
+import org.firstinspires.ftc.robotcore.external.Telemetry;
 import org.firstinspires.ftc.teamcode.CenterStageRobot.subsystems.IntakeArmSubsystem;
 import org.firstinspires.ftc.teamcode.CenterStageRobot.subsystems.IntakeSubsystem;
 import org.firstinspires.ftc.teamcode.CenterStageRobot.subsystems.OuttakeSusystem;
@@ -12,35 +13,24 @@ import org.firstinspires.ftc.teamcode.CenterStageRobot.subsystems.PixelColorDete
 
 public class IntakeCommand extends CommandBase {
     private IntakeSubsystem intakeSubsystem;
-    private IntakeArmSubsystem intakeArmSubsystem;
-    private OuttakeSusystem outtakeSusystem;
     private PixelColorDetectorSubsystem pixelColorDetectorSubsystem;
-    public IntakeCommand(IntakeSubsystem intakeSubsystem, PixelColorDetectorSubsystem pixelColorDetectorSubsystem,
-                         OuttakeSusystem outtakeSusystem, IntakeArmSubsystem intakeArmSubsystem) {
+
+    private Telemetry telemetry;
+    public IntakeCommand(IntakeSubsystem intakeSubsystem, PixelColorDetectorSubsystem pixelColorDetectorSubsystem, Telemetry telemetry) {
         this.intakeSubsystem = intakeSubsystem;
         this.pixelColorDetectorSubsystem = pixelColorDetectorSubsystem;
-        this.intakeArmSubsystem = intakeArmSubsystem;
-        this.outtakeSusystem = outtakeSusystem;
 
-        addRequirements(intakeSubsystem, pixelColorDetectorSubsystem, outtakeSusystem, intakeArmSubsystem);
+        this.telemetry = telemetry;
+
+
+        telemetry.addData("Intake Init", "");
+
+        addRequirements(intakeSubsystem, pixelColorDetectorSubsystem);
     }
 
     @Override
     public void initialize() {
         intakeSubsystem.run();
-    }
-
-    @Override
-    public void end(boolean interrupted) {
-        new SequentialCommandGroup(
-                new InstantCommand(outtakeSusystem::wheel_stop),
-                new InstantCommand(intakeArmSubsystem::raiseArm),
-                new WaitCommand(200),
-                new InstantCommand(intakeSubsystem::reverse, intakeSubsystem),
-                new WaitCommand(600),
-                new InstantCommand(intakeSubsystem::stop, intakeSubsystem)
-        ).schedule();
-//        intakeSubsystem.stop();
     }
 
     @Override
