+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/CenterStageRobot/subsystems/IntakeSubsystem.java
@@ -60,6 +60,8 @@ public class IntakeSubsystem extends SubsystemBase {
 
     public void run() {
         state = State.LOADING;
+
+        telemetry.addData("Intake Running", "");
         motor.set(speed);
     }
